+++ app/tests/TestCase.php
@@ -12,7 +12,6 @@ namespace UserFrosting\Tests;
 
 use PHPUnit\Framework\TestCase as BaseTestCase;
 use UserFrosting\System\UserFrosting;
-use UserFrosting\Sprinkle\Core\Facades\Debug;
 
 /**
  * Class to handle Test
@@ -58,28 +57,11 @@ class TestCase extends BaseTestCase
             $this->refreshApplication();
         }
 
-        $this->setUpTraits();
-
         foreach ($this->afterApplicationCreatedCallbacks as $callback) {
             call_user_func($callback);
         }
 
         $this->setUpHasRun = true;
-    }
-
-    /**
-     * Boot the testing helper traits.
-     *
-     * @deprecated
-     */
-    protected function setUpTraits()
-    {
-        $uses = array_flip(class_uses_recursive(static::class));
-
-        if (isset($uses[DatabaseTransactions::class])) {
-            Debug::warning("`UserFrosting\Tests\DatabaseTransactions` has been deprecated and will be removed in future versions.  Please use `UserFrosting\Sprinkle\Core\Tests\DatabaseTransactions` class instead.");
-            $this->beginDatabaseTransaction();
-        }
     }
 
     /**