+++ CHANGELOG.md
@@ -14,13 +14,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 ### Changed
-- Changed `.php_cs` to `.php_cs.dist`
-- Changed `phpunit.xml` to `phpunit.xml.dist`
+- Changed `.php_cs` to `.php_cs.dist`.
+- Changed `phpunit.xml` to `phpunit.xml.dist`.
+- Moved `*Command` Bakery class to `Command/*` namespace.
+- Replaced `UserFrosting\System\Bakery\BaseCommand` with `UserFrosting\Sprinkle\Core\Bakery\AbstractCommand`.
 
 ### Fixed
 
 ### Removed
-- Removed deprecated `UserFrosting\System\Bakery\Migration` from 4.2.0.
+- Removed deprecated `UserFrosting\System\Bakery\Migration` (deprecated in 4.2.0).
+- Removed deprecated `UserFrosting\Tests\DatabaseTransactions` (deprecated in 4.2.0).
 
 ## [v4.4.1]